Refactor player profile code into separate module and clean imports
